+++ src/Field/Types/Color.php
@@ -1,6 +1,6 @@
-<?php namespace Frozennode\Administrator\Fields\Types;
+<?php namespace Frozennode\Administrator\Field\Types;
 
-class DateTime extends Time {
+class Color extends Text {
 
 	/**
 	 * Abstract method that should return a field's string representation in the config files
@@ -9,6 +9,7 @@ class DateTime extends Time {
 	 */
 	public function getConfigName()
 	{
-		return 'datetime';
+		return 'color';
 	}
+
 }
